requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
openai>=0.27.0
python-dotenv>=0.19.0
PyMuPDF>=1.18.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pydantic>=1.8.0